requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
flask-cors==3.1.1
deepface==0.0.78
opencv-python==4.7.0.72
Pillow==9.4.0
